Simplification: SDK Surface Audit 3 (#543)
* more explicit exports pt.1

* explicit package exports

* Drop internal import from index

* Move and export token transfer types

* export errors as types
