--- packages/client/wallets/aa/src/api/index.ts
@@ -1,1 +1,0 @@
-export * from "./CrossmintWalletService";
