--- packages/client/wallets/aa/src/blockchain/index.ts
@@ -1,4 +1,0 @@
-export * from "./token";
-export * from "./wallets";
-
-export * from "./BlockchainNetworks";
