+++ packages/client/wallets/aa/src/blockchain/transfer.ts
@@ -1,12 +1,7 @@
 import { Abi, Account, Address, erc20Abi, erc721Abi } from "viem";
 
-import erc1155Abi from "../../ABI/ERC1155.json";
-import { ERC2OEVMToken, NFTEVMToken, SFTEVMToken } from "./Tokens";
-
-export type ERC20TransferType = { token: ERC2OEVMToken; amount: bigint };
-export type SFTTransferType = { token: SFTEVMToken; quantity: number };
-export type NFTTransferType = { token: NFTEVMToken };
-export type TransferType = ERC20TransferType | SFTTransferType | NFTTransferType;
+import erc1155Abi from "../ABI/ERC1155.json";
+import type { ERC20TransferType, SFTTransferType, TransferType } from "../types/Tokens";
 
 type TransferInputParams = {
     from: Account;
