+++ packages/client/wallets/aa/src/blockchain/wallets/passkey.ts
@@ -1,13 +1,6 @@
-import {
-    CrossmintWalletService,
-    EVMSmartWallet,
-    EntryPointDetails,
-    PasskeySigner,
-    UserParams,
-    WalletConfig,
-    blockchainToChainId,
-} from "@/index";
+import { CrossmintWalletService } from "@/api/CrossmintWalletService";
 import { PasskeySignerData } from "@/types/API";
+import { EntryPointDetails, PasskeySigner, UserParams, WalletConfig } from "@/types/Config";
 import { WalletCreationParams } from "@/types/internal";
 import { CURRENT_VERSION, ZERO_DEV_TYPE } from "@/utils/constants";
 import { createPasskeyValidator, deserializePasskeyValidator } from "@zerodev/passkey-validator";
@@ -15,7 +8,10 @@ import { KernelValidator, createKernelAccount } from "@zerodev/sdk";
 import { EntryPoint } from "permissionless/types/entrypoint";
 import { PublicClient } from "viem";
 
+import { blockchainToChainId } from "@crossmint/common-sdk-base";
+
 import { deserializePasskeyValidatorData, serializePasskeyValidatorData } from "../../utils/passkey";
+import { EVMSmartWallet } from "./EVMSmartWallet";
 
 export interface PasskeyWalletParams extends WalletCreationParams {
     walletConfig: WalletConfig & { signer: PasskeySigner };
