+++ packages/client/wallets/aa/src/index.ts
@@ -1,7 +1,26 @@
 export { EVMBlockchainIncludingTestnet as Blockchain, blockchainToChainId } from "@crossmint/common-sdk-base";
 
-export * from "./api";
-export * from "./blockchain";
-export * from "./types";
+export { EVMSmartWallet } from "./blockchain/wallets/EVMSmartWallet";
 
-export * from "./SmartWalletSDK";
+export type {
+    SmartWalletSDKInitParams,
+    UserParams,
+    Web3AuthSigner,
+    ViemAccount,
+    PasskeySigner,
+    EOASigner,
+    WalletConfig,
+} from "./types/Config";
+
+export type { TransferType, ERC20TransferType, NFTTransferType, SFTTransferType } from "./types/Tokens";
+
+export type {
+    NotAuthorizedError,
+    TransferError,
+    TransactionError,
+    CrossmintServiceError,
+    RunningOnServerError,
+    WalletSdkError,
+} from "./types/Error";
+
+export { SmartWalletSDK } from "./SmartWalletSDK";
