+++ packages/client/wallets/aa/src/types/Error.ts
@@ -1,5 +1,3 @@
-import { logError } from "@/services/logging";
-
 export class NotAuthorizedError extends Error {
     code = "ERROR_NOT_AUTHORIZED";
 
@@ -56,15 +54,4 @@ export class WalletSdkError extends Error {
         super(message);
         Object.setPrototypeOf(this, WalletSdkError.prototype);
     }
-}
-
-export function errorToJSON(error: Error | unknown) {
-    const errorToLog = error instanceof Error ? error : { message: "Unknown error", name: "Unknown error" };
-
-    if (!(errorToLog instanceof Error) && (errorToLog as any).constructor?.name !== "SyntheticBaseEvent") {
-        logError("ERROR_TO_JSON_FAILED", { error: errorToLog });
-        throw new Error("[errorToJSON] err is not instanceof Error nor SyntheticBaseEvent");
-    }
-
-    return JSON.parse(JSON.stringify(errorToLog, Object.getOwnPropertyNames(errorToLog)));
 }
