+++ packages/client/wallets/aa/src/types/Tokens.ts
@@ -20,3 +20,8 @@ export interface ERC2OEVMToken extends EVMToken {
 }
 
 export type TokenType = "nft" | "sft" | "ft";
+
+export type ERC20TransferType = { token: ERC2OEVMToken; amount: bigint };
+export type SFTTransferType = { token: SFTEVMToken; quantity: number };
+export type NFTTransferType = { token: NFTEVMToken };
+export type TransferType = ERC20TransferType | SFTTransferType | NFTTransferType;
