--- packages/client/wallets/aa/src/types/index.ts
@@ -1,2 +1,0 @@
-export * from "./Config";
-export * from "./Error";
